telnyx-twilio-migration/migration-examples/python-flask/clicktocall-flask/requirements.txt
Flask==1.1.2
tox==3.14.6
telnyx>=2.0,<3.0
python-dotenv==0.12.0
